feedback

+++ src/java/org/apache/cassandra/exceptions/TombstoneAbortException.java
@@ -30,7 +30,7 @@ public class TombstoneAbortException extends ReadAbortException
     public final int nodes;
     public final long tombstones;
 
-    public TombstoneAbortException(int nodes, int tombstones, String cql, boolean dataPresent, ConsistencyLevel consistency, int received, int blockFor, Map<InetAddressAndPort, RequestFailureReason> failureReasonByEndpoint)
+    public TombstoneAbortException(int nodes, long tombstones, String cql, boolean dataPresent, ConsistencyLevel consistency, int received, int blockFor, Map<InetAddressAndPort, RequestFailureReason> failureReasonByEndpoint)
     {
         super(tombstoneAbortMessage(nodes, tombstones, cql), consistency, received, blockFor, dataPresent, failureReasonByEndpoint);
         this.nodes = nodes;
